Integrate OpenTracing to the project.
Global tracer is registered (the host is located at localhost:5775 and hardwired that
way for now).
Also, standalone.xml and a custom-made module is copied out into this project as some
of the configuration is specific to the project and the default standalone.xml
wouldn't work with it (especially when it comes to datasources).